django-rest-framework/tutorial/requirements.txt
Pygments==2.2.0
alauda_django_oauth==0.9.0
djangorestframework==3.6.3
django==1.11.4
django-cors-middleware==1.3.1
django-oauth-toolkit==1.0.0
coreapi==2.0.0
coreschema==0.0.4
oauthlib==2.0.1

